VehicleServive and VehicleServiceImpl has completed. Let me know if we need the addVehicle obj. Fekadu

+++ TrafficAccidentsAnalysisBackEnd/src/main/java/com/trafficaccidentsanalysis/backend/repository/VehicleRepository.java
@@ -4,6 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.trafficaccidentsanalysis.backend.model.Vehicle;
 
-public interface VehicleRepository extends JpaRepository<Vehicle, Long> {
+public interface VehicleRepository extends JpaRepository<Vehicle, Integer> {
+
 
 }
